Update issue 254 Now depth and stencil could be updated seperately. Prototype code for StencilMirror. Add mat_reflect to eflib.

+++ salviax/include/resource/mesh/sa/skin_mesh_impl.h
@@ -78,7 +78,7 @@ private:
 	{
 		// Find segment
 		std::vector<float>::iterator it = lower_bound(aninfo->X.begin(), aninfo->X.end(), time);
-		int i_segment = distance(aninfo->X.begin(), it) - 1;
+		size_t i_segment = distance(aninfo->X.begin(), it) - 1;
 
 		if( it == aninfo->X.begin() )
 		{
